+++ 2017/Oct 2025 c++/Day 1 Inverse Captcha/Day1.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void Day1PartOne();
+void Day1PartTwo();
